chore: tests for server
