+++ package.json
@@ -16,13 +16,13 @@
     "heroku-postbuild": "yarn build"
   },
   "devDependencies": {
-    "ava": "^1.2.1",
     "@commitlint/cli": "^7.3.2",
     "@commitlint/config-conventional": "^7.3.1",
     "@types/express": "^4.16.0",
     "@types/node": "^10.12.18",
     "@types/uuid": "^3.4.4",
     "@types/ws": "^6.0.1",
+    "ava": "^1.2.1",
     "ts-node": "^8.0.2",
     "tslint": "^5.12.1",
     "tslint-eslint-rules": "^5.4.0",
@@ -50,10 +50,16 @@
     }
   },
   "ava": {
+    "extensions": [
+      "ts"
+    ],
+    "require": [
+      "ts-node/register"
+    ],
     "files": [
-      "test/**/*.js"
+      "test/**/*"
     ]
-},
+  },
   "commitlint": {
     "extends": [
       "@commitlint/config-conventional"
